requirements.txt
django-autoslug
lxml
Unidecode
